Prod incident: * Loop over stages once rather than twice in updateStages() * Investigate using List instead of Set in StageService for high CPU usage
