+++ docker-compose.yml
@@ -3,7 +3,7 @@ version: '3.1'
 services:
 
   postgres:
-    image: postgres
+    image: postgres:9.6
     restart: always
     ports:
       - 5432:5432
