--- src/main/java/uk/gov/digital/ho/hocs/casework/api/WorkingDaysElapsedProvider.java
@@ -1,8 +1,0 @@
-package uk.gov.digital.ho.hocs.casework.api;
-
-import java.time.LocalDate;
-
-public interface WorkingDaysElapsedProvider {
-
-    Integer getWorkingDaysSince(String caseType, LocalDate fromDate);
-}
