+++ src/test/java/uk/gov/digital/ho/hocs/casework/priority/policy/WorkingDaysElapsedPolicyTest.java
@@ -5,7 +5,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
-import uk.gov.digital.ho.hocs.casework.api.WorkingDaysElapsedProvider;
+import uk.gov.digital.ho.hocs.casework.client.infoclient.InfoClient;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
@@ -30,34 +30,34 @@ public class WorkingDaysElapsedPolicyTest {
     private static final String TEST_CASE_TYPE = "CASE_TYPE_1";
 
     @Mock
-    private WorkingDaysElapsedProvider workingDaysElapsedProvider;
+    private InfoClient infoClient;
 
     @Before
     public void before() {
-        policy = new WorkingDaysElapsedPolicy(workingDaysElapsedProvider, PROPERTY_NAME, PROPERTY_VALUE, DATE_FIELD_NAME, DATE_FORMAT, CAP_NUMBER_OF_DAYS,
-                CAP_POINTS_TO_AWARD, POINTS_TO_AWARD_PER_DAY);
+        policy = new WorkingDaysElapsedPolicy(PROPERTY_NAME, PROPERTY_VALUE, DATE_FIELD_NAME, DATE_FORMAT, CAP_NUMBER_OF_DAYS,
+                CAP_POINTS_TO_AWARD, POINTS_TO_AWARD_PER_DAY, infoClient);
     }
 
     @Test
     public void apply_criteriaMatched() {
         LocalDate testDate = LocalDate.now().minusDays(10);
-        when(workingDaysElapsedProvider.getWorkingDaysSince(TEST_CASE_TYPE, testDate)).thenReturn(10);
+        when(infoClient.getWorkingDaysElapsedForCaseType(TEST_CASE_TYPE, testDate)).thenReturn(10);
         double result = policy.apply(Map.of(CASE_TYPE, TEST_CASE_TYPE, PROPERTY_NAME, PROPERTY_VALUE, DATE_FIELD_NAME, DateTimeFormatter.ofPattern(DATE_FORMAT).format(testDate)));
         assertThat(result).isEqualTo(20d);
 
-        verify(workingDaysElapsedProvider).getWorkingDaysSince(TEST_CASE_TYPE, testDate);
-        verifyNoMoreInteractions(workingDaysElapsedProvider);
+        verify(infoClient).getWorkingDaysElapsedForCaseType(TEST_CASE_TYPE, testDate);
+        verifyNoMoreInteractions(infoClient);
     }
 
     @Test
     public void apply_criteriaMatched_capped() {
         LocalDate testDate = LocalDate.now().minusDays(74);
-        when(workingDaysElapsedProvider.getWorkingDaysSince(TEST_CASE_TYPE, testDate)).thenReturn(55);
+        when(infoClient.getWorkingDaysElapsedForCaseType(TEST_CASE_TYPE, testDate)).thenReturn(55);
         double result = policy.apply(Map.of(CASE_TYPE, TEST_CASE_TYPE, PROPERTY_NAME, PROPERTY_VALUE, DATE_FIELD_NAME, DateTimeFormatter.ofPattern(DATE_FORMAT).format(testDate)));
         assertThat(result).isEqualTo(35d);
 
-        verify(workingDaysElapsedProvider).getWorkingDaysSince(TEST_CASE_TYPE, testDate);
-        verifyNoMoreInteractions(workingDaysElapsedProvider);
+        verify(infoClient).getWorkingDaysElapsedForCaseType(TEST_CASE_TYPE, testDate);
+        verifyNoMoreInteractions(infoClient);
     }
 
     @Test
@@ -65,7 +65,7 @@ public class WorkingDaysElapsedPolicyTest {
         double result = policy.apply(Map.of(PROPERTY_NAME, "C"));
         assertThat(result).isEqualTo(0);
 
-        verifyNoMoreInteractions(workingDaysElapsedProvider);
+        verifyNoMoreInteractions(infoClient);
     }
 
     @Test
@@ -73,6 +73,6 @@ public class WorkingDaysElapsedPolicyTest {
         double result = policy.apply(Map.of());
         assertThat(result).isEqualTo(0);
 
-        verifyNoMoreInteractions(workingDaysElapsedProvider);
+        verifyNoMoreInteractions(infoClient);
     }
 }
